feat: create a FeedbackEnvironmentObserver
