--- Sources/FeedbackBulb/Zofx.swift
@@ -1,6 +1,0 @@
-public struct FeedbackBulb {
-  public private(set) var text = "Hello, World!"
-
-  public init() {
-  }
-}
